+++ flutter_ui_bloc/lib/src/form/validation/validation_translations.dart
@@ -47,6 +47,8 @@ abstract class ValidationTranslations {
     this.formats = const ValidationFormats(),
   });
 
+  String invalid();
+
   String greaterThanComparable(Object value);
 
   String lessThanComparable(Object value);
@@ -55,9 +57,17 @@ abstract class ValidationTranslations {
 
   String lessOrEqualThanComparable(Object value);
 
-  String beforeDateTime(Object value);
+  String minLength(int value);
 
-  String afterDateTime(Object value);
+  String maxLength(int value);
+
+  String match(String value);
+
+  String notMatch(String value);
+
+  String beforeDateTime(DateTime value);
+
+  String afterDateTime(DateTime value);
 
   ValidationTranslations copyWith({ValidationFormats? formats});
 
@@ -75,34 +85,39 @@ class ValidationEnTranslations extends ValidationTranslations {
   }) : super(formats: formats);
 
   @override
-  String greaterThanComparable(Object value) {
-    return 'It must be greater than ${formats.format(value)}.';
-  }
+  String invalid() => 'It not valid.';
 
   @override
-  String lessThanComparable(Object value) {
-    return 'It must be less than ${formats.format(value)}.';
-  }
+  String greaterThanComparable(Object value) => 'It must be greater than ${formats.format(value)}.';
 
   @override
-  String greaterOrEqualThanComparable(Object value) {
-    return 'It must be greater or equal than ${formats.format(value)}.';
-  }
+  String lessThanComparable(Object value) => 'It must be less than ${formats.format(value)}.';
 
   @override
-  String lessOrEqualThanComparable(Object value) {
-    return 'It must be less or equal than ${formats.format(value)}.';
-  }
+  String greaterOrEqualThanComparable(Object value) =>
+      'It must be greater or equal than ${formats.format(value)}.';
 
   @override
-  String beforeDateTime(Object value) {
-    return 'It must be before ${formats.format(value)}.';
-  }
+  String lessOrEqualThanComparable(Object value) =>
+      'It must be less or equal than ${formats.format(value)}.';
 
   @override
-  String afterDateTime(Object value) {
-    return 'It must be after ${formats.format(value)}.';
-  }
+  String minLength(int value) => 'It must be at least $value characters.';
+
+  @override
+  String maxLength(int value) => 'It must be at most $value characters.';
+
+  @override
+  String match(String value) => invalid();
+
+  @override
+  String notMatch(String value) => invalid();
+
+  @override
+  String beforeDateTime(DateTime value) => 'It must be before ${formats.format(value)}.';
+
+  @override
+  String afterDateTime(DateTime value) => 'It must be after ${formats.format(value)}.';
 
   @override
   ValidationTranslations copyWith({ValidationFormats? formats}) {
@@ -119,6 +134,33 @@ class ValidationItTranslations extends ValidationEnTranslations {
   const ValidationItTranslations({
     ValidationFormats formats = const ValidationFormats(),
   }) : super(formats: formats);
+
+  @override
+  String invalid() => 'Non è valido.';
+
+  @override
+  String greaterThanComparable(Object value) => 'Maggiore di ${formats.format(value)}.';
+
+  @override
+  String lessThanComparable(Object value) => 'Minore di ${formats.format(value)}.';
+
+  @override
+  String greaterOrEqualThanComparable(Object value) => 'Maggiore o uguale di ${formats.format(value)}.';
+
+  @override
+  String lessOrEqualThanComparable(Object value) => 'Minore o uguale di ${formats.format(value)}.';
+
+  @override
+  String minLength(int value) => 'Almeno $value caratteri.';
+
+  @override
+  String maxLength(int value) => 'Massimo $value caratteri.';
+
+  @override
+  String beforeDateTime(DateTime value) => 'Prima di ${formats.format(value)}.';
+
+  @override
+  String afterDateTime(DateTime value) => 'Dopo di ${formats.format(value)}.';
 
   @override
   ValidationTranslations copyWith({ValidationFormats? formats}) {